+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/service/DaoOnlyEntireLookupService.java
@@ -92,6 +92,18 @@ public class DaoOnlyEntireLookupService<E extends Entity<?>> implements EntireLo
         }
     }
 
+    /**
+     * @since 1.4.1
+     */
+    @Override
+    public int lookupCount() throws ServiceException {
+        try {
+            return dao.lookupCount();
+        } catch (Exception e) {
+            throw ServiceExceptionHelper.logAndThrow("查询全部实体数量时发生异常", exceptionLogLevel, sem, e);
+        }
+    }
+
     public EntireLookupDao<E> getDao() {
         return dao;
     }